django_rest_todos/requirements.txt
Django==2.0.7
djangorestframework==3.8.2
pytz==2018.5
shell==1.0.1
